+++ src/componentes/escolas/GeracaoDaAta/ModalReverDevolucoesAoTesouro.js
@@ -1,0 +1,19 @@
+import React from "react";
+import {ModalBootstrap} from "../../Globais/ModalBootstrap";
+
+export const ModalReverDevolucoesAoTesouro = (props)=> {
+  return(
+      <ModalBootstrap
+          show={props.show}
+          onHide={props.handleClose}
+          titulo={props.titulo}
+          bodyText={props.texto}
+          primeiroBotaoOnclick={()=>props.setShowModalDevolucoesAoTesouro(true)}
+          primeiroBotaoTexto={props.primeiroBotaoTexto}
+          primeiroBotaoCss={props.primeiroBotaoCss}
+          segundoBotaoOnclick={props.handleClose}
+          segundoBotaoCss={props.segundoBotaoCss}
+          segundoBotaoTexto={props.segundoBotaoTexto}
+      />
+  )
+};
